+++ views/single-post.handlebars
@@ -4,13 +4,17 @@
   <div>
     <textarea name="comment-body"></textarea>
   </div>
-
+  
   <div>
     <button type="submit">add comment</button>
   </div>
 </form>
+<div>
+    <button type="submit">Send Email</button>
+</div>
 {{/if}}
 {{> comments post.comments}}
 {{#if loggedIn}}
 <script src="/js/comment.js"></script>
+<script src="/js/email.js"></script>
 {{/if}}
